requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
git-filter-repo==2.47.0
idna==3.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
